Add ability to add a custom initializer.
This is needed in cases where a particle system's
initializer needs to be updated on the fly (such as from
a sensor listener). Prior to this a new ParticleSystem
would have to be created or extra initializers would have
to be added. Since the extra initializers simply override
the previous initialization, they bloat memory.

Fixes #75